Test Scene JAM
Added a test scene for user JAM. Fiddling around with the core mechanics. Made a test player script with auto movement and jump implemented. As well as a camera follow scrip. Attempting to work at sliding and enemies. All of it is labeled test as we're still not sure which method we want to go for.