Prototype Scene
Added a prototype scene and begun work on the alpha build. Reworked the controls on a new and official Player Controller. Added touch controls . Autorunning works, jumping on tap works (but will be reworked) dash works (but haven't been implemented in a control system yet) and swiping added for slide but slide has not yet been worked on. Minor changes to camera and aspect.