Collision and Tiles
After many attempts, I've finally fixed composite collider. I figured out that the velocity of the player and the strength of the collider lead to the player penetrating the collider. Fixed it by adjusting the physics of the composite collider so that now it's almost impenetrable. Also added 2 new tile palettes.