Attack Rework
Completely revamped the Attack move. Instead of snapping to targets it now move towards it in a set time. The bounce is now consistent from each enemy kill. Animations added to everything so that the bean looks angry as he flies towards the target and then does a little spin after killing the enemy. Everything works fluently. Fixed an attack bug as well that lead to players sometimes not being able to attack. This was due to the layers of the pickups, and this has been changed. Attack is now more or less 100% consistent.